+++ cli/src/main.rs
@@ -21,8 +21,8 @@ fn main() {
     let config = Configuration::from(cli.clone());
 
     if let Some(source) = cli.source {
-        let working_dir = std::env::current_dir().expect("No working dir set");
-        exit(!run(source, working_dir, config) as i32)
+        let source_dir = std::env::current_dir().expect("No working dir set");
+        exit(!run(source, source_dir, config) as i32)
     }
     repl(config);
 }
